Break out fakeLocation into own file
