+++ cypress/support/commands.js
@@ -11,24 +11,8 @@
 //
 // -- This is a parent command --
 // Cypress.Commands.add("login", (email, password) => { ... })
-// thanks to https://github.com/cypress-io/cypress/issues/2671
-function fakeLocation(latitude, longitude) {
-  return {
-    onBeforeLoad(win) {
-      cy.stub(win.navigator.geolocation, 'getCurrentPosition', (cb, err) => {
-        if (latitude && longitude) {
-          return cb({
-            coords: {
-              latitude,
-              longitude,
-            },
-          });
-        }
-        throw err({ code: 1 }); // 1: rejected, 2: unable, 3: timeout
-      });
-    },
-  };
-}
+import fakeLocation from './helpers/fakeLocation';
+
 // sets up the /cities landing page tests
 Cypress.Commands.add('seedCitiesPage', () => {
   cy.request({
